E0-10: размещение по пакетам
