--- app/src/main/java/ru/practicum/android/diploma/data/dto/SkillDto.kt
@@ -1,3 +1,0 @@
-package ru.practicum.android.diploma.data.dto
-
-data class SkillDto(val name: String)
