+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/CountryDto.kt
@@ -1,3 +1,3 @@
-package ru.practicum.android.diploma.data.dto
+package ru.practicum.android.diploma.data.search.dto
 
 data class CountryDto(val id: String, val name: String)
