+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/ExperienceDto.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.dto
+package ru.practicum.android.diploma.data.search.dto
 
 data class ExperienceDto(
     val id: String,
