+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/IndustryDto.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.dto
+package ru.practicum.android.diploma.data.search.dto
 
 data class IndustryDto(
     val id:String,
