+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/SalaryDto.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.dto
+package ru.practicum.android.diploma.data.search.dto
 
 data class SalaryDto(
     val from: Int? = null,
