+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/ScheduleDto.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.dto
+package ru.practicum.android.diploma.data.search.dto
 
 data class ScheduleDto(
     val id: String,
