+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/SkillDto.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.data.search.dto
+
+data class SkillDto(val name: String)
