+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/VacancyDetailedDto.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.dto
+package ru.practicum.android.diploma.data.search.dto
 
 import com.google.gson.annotations.SerializedName
