+++ app/src/main/java/ru/practicum/android/diploma/domain/search/model/SearchVacancyOptions.kt
@@ -1,0 +1,10 @@
+package ru.practicum.android.diploma.domain.search.model
+
+class SearchVacancyOptions(
+    var text: String,
+    var page: Int = 0,
+    var areaId: String? = null,
+    var industryId: String? = null,
+    var salary: Int? = null,
+    var onlyWithSalary: Boolean? = null,
+)
